LLMs/Task_3_AI_Resume_Screening_System/Task_3_AI_Resume_Screening_System/requirements.txt
# ─────────────────────────────────────────────────────────────
# AI Resume Screening System — Requirements
# Task 3 | Innomatics Research Labs
#
# Install with:
#   pip install -r requirements.txt
#
# Versions are intentionally unpinned to let pip resolve
# compatible versions automatically and avoid conflicts.
# ─────────────────────────────────────────────────────────────

# Core LangChain (resolves to ~1.x)
langchain
langchain-openai
langchain-community

# LangSmith for tracing
langsmith

# OpenAI SDK
openai

# Environment management
python-dotenv

# Streamlit UI
streamlit

# Utilities
pydantic
